feat: add alarm management components and hooks
- Introduced AlarmCard component for displaying individual alarms with edit and delete functionality.
- Added AlarmEditor component for creating and editing alarms, including time, label, repeat options, and ringtone selection.
- Implemented WheelPicker component for a user-friendly time selection interface.
- Created useAlarmState hook for centralized state management of alarms, including adding, editing, deleting, and toggling alarms.
- Developed utility functions for alarm persistence, normalization, and formatting.
- Added types for alarm items and wheel picker props to enhance type safety.
- Established constants for wheel picker configuration.
