+++ src/renderer/components/states/maxExpand/components/alarm/components/AlarmTab.tsx
@@ -24,182 +24,12 @@
  * @author 鸡哥
  */
 
-import React, { useState, useEffect, useCallback, useRef } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { SvgIcon } from '../../../../../../utils/SvgIcon';
-import {
-  DEFAULT_SYSTEM_ALARM_RINGTONE,
-  normalizeSystemAlarmRingtone,
-  previewAlarmSound,
-  stopPreviewAlarmSound,
-  subscribePreviewAlarmSoundState,
-  SYSTEM_ALARM_RINGTONE_OPTIONS,
-  type SystemAlarmRingtone,
-} from '../../../../../../utils/audio/alarmSound';
-
-/** 星期几 0=周日 ... 6=周六 */
-type Weekday = 0 | 1 | 2 | 3 | 4 | 5 | 6;
-
-/** 单条闹钟 */
-interface AlarmItem {
-  id: number;
-  hour: number;
-  minute: number;
-  second: number;
-  label: string;
-  enabled: boolean;
-  repeat: Weekday[];
-  ringtone: SystemAlarmRingtone;
-  loop: boolean;
-  createdAt: number;
-}
-
-const STORE_KEY = 'alarms';
-
-const ALL_WEEKDAYS: Weekday[] = [1, 2, 3, 4, 5, 6, 0];
-
-/** 通过 IPC 写入文件 */
-function persistAlarms(items: AlarmItem[]): void {
-  window.api.storeWrite(STORE_KEY, items).catch(() => {});
-}
-
-/** 规范化旧数据 */
-function normalizeAlarms(items: AlarmItem[]): AlarmItem[] {
-  return items.map((a) => ({
-    ...a,
-    hour: a.hour ?? 0,
-    minute: a.minute ?? 0,
-    second: a.second ?? 0,
-    label: a.label ?? '',
-    enabled: a.enabled ?? true,
-    repeat: Array.isArray(a.repeat) ? a.repeat : [],
-    ringtone: normalizeSystemAlarmRingtone(a.ringtone),
-    loop: typeof a.loop === 'boolean' ? a.loop : true,
-    createdAt: a.createdAt ?? Date.now(),
-  }));
-}
-
-/** 格式化时间 HH:MM:SS */
-function formatTime(h: number, m: number, s: number): string {
-  return `${String(h).padStart(2, '0')}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`;
-}
-
-/** iOS 风格时间轮盘选择器 */
-const WHEEL_ITEM_HEIGHT = 28;
-const WHEEL_VISIBLE_ITEMS = 5;
-
-interface WheelPickerProps {
-  min: number;
-  max: number;
-  value: number;
-  onChange: (val: number) => void;
-}
-
-function WheelPicker({ min, max, value, onChange }: WheelPickerProps): React.ReactElement {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const items = React.useMemo(() => Array.from({ length: max - min + 1 }, (_, i) => min + i), [min, max]);
-  const padding = Math.floor(WHEEL_VISIBLE_ITEMS / 2);
-  const containerHeight = WHEEL_VISIBLE_ITEMS * WHEEL_ITEM_HEIGHT;
-
-  const currentIdx = value - min;
-  const dragStartY = useRef(0);
-  const dragStartIdx = useRef(0);
-  const isDragging = useRef(false);
-  const lastIdxRef = useRef(currentIdx);
-  lastIdxRef.current = currentIdx;
-
-  const clampIdx = useCallback((idx: number) => Math.max(0, Math.min(items.length - 1, idx)), [items.length]);
-
-  const scrollToIdx = useCallback((idx: number, smooth = true) => {
-    const el = containerRef.current;
-    if (!el) return;
-    const top = idx * WHEEL_ITEM_HEIGHT;
-    if (smooth) {
-      el.scrollTo({ top, behavior: 'smooth' });
-    } else {
-      el.scrollTop = top;
-    }
-  }, []);
-
-  useEffect(() => {
-    if (!isDragging.current) {
-      scrollToIdx(currentIdx, false);
-    }
-  }, [currentIdx, scrollToIdx]);
-
-  useEffect(() => {
-    const el = containerRef.current;
-    if (!el) return;
-    const handleWheel = (e: WheelEvent): void => {
-      e.preventDefault();
-      const dir = e.deltaY > 0 ? 1 : e.deltaY < 0 ? -1 : 0;
-      if (dir === 0) return;
-      const cur = lastIdxRef.current;
-      const next = clampIdx(cur + dir);
-      if (next !== cur) {
-        scrollToIdx(next, true);
-        onChange(items[next]);
-      }
-    };
-    el.addEventListener('wheel', handleWheel, { passive: false });
-    return () => el.removeEventListener('wheel', handleWheel);
-  }, [items, onChange, clampIdx, scrollToIdx]);
-
-  const handleMouseDown = useCallback((e: React.MouseEvent) => {
-    e.preventDefault();
-    isDragging.current = true;
-    dragStartY.current = e.clientY;
-    dragStartIdx.current = lastIdxRef.current;
-
-    const handleMouseMove = (ev: MouseEvent): void => {
-      const dy = dragStartY.current - ev.clientY;
-      const offsetItems = Math.round(dy / WHEEL_ITEM_HEIGHT);
-      const next = clampIdx(dragStartIdx.current + offsetItems);
-      if (next !== lastIdxRef.current) {
-        scrollToIdx(next, false);
-        onChange(items[next]);
-      }
-    };
-
-    const handleMouseUp = (): void => {
-      isDragging.current = false;
-      document.removeEventListener('mousemove', handleMouseMove);
-      document.removeEventListener('mouseup', handleMouseUp);
-      scrollToIdx(lastIdxRef.current, true);
-    };
-
-    document.addEventListener('mousemove', handleMouseMove);
-    document.addEventListener('mouseup', handleMouseUp);
-  }, [items, onChange, clampIdx, scrollToIdx]);
-
-  return (
-    <div className="alarm-wheel-picker" style={{ height: containerHeight }}>
-      <div className="alarm-wheel-highlight" />
-      <div
-        className="alarm-wheel-scroll"
-        ref={containerRef}
-        onMouseDown={handleMouseDown}
-        style={{ height: containerHeight, cursor: 'grab' }}
-      >
-        {Array.from({ length: padding }, (_, i) => (
-          <div key={`pt${i}`} className="alarm-wheel-item alarm-wheel-item--empty" style={{ height: WHEEL_ITEM_HEIGHT }} />
-        ))}
-        {items.map((n) => (
-          <div
-            key={n}
-            className={`alarm-wheel-item${n === value ? ' alarm-wheel-item--active' : ''}`}
-            style={{ height: WHEEL_ITEM_HEIGHT }}
-          >
-            {String(n).padStart(2, '0')}
-          </div>
-        ))}
-        {Array.from({ length: padding }, (_, i) => (
-          <div key={`pb${i}`} className="alarm-wheel-item alarm-wheel-item--empty" style={{ height: WHEEL_ITEM_HEIGHT }} />
-        ))}
-      </div>
-    </div>
-  );
-}
+import { useAlarmState } from '../hooks/useAlarmState';
+import { AlarmCard } from './AlarmCard';
+import { AlarmEditor } from './AlarmEditor';
 
 /**
  * Alarm Tab
@@ -207,259 +37,22 @@ function WheelPicker({ min, max, value, onChange }: WheelPickerProps): React.Rea
  */
 export function AlarmTab(): React.ReactElement {
   const { t } = useTranslation();
-  const [alarms, setAlarms] = useState<AlarmItem[]>([]);
-  const [loaded, setLoaded] = useState(false);
-  const skipPersistOnceRef = useRef(false);
-  const [previewPlaying, setPreviewPlaying] = useState(false);
+  const state = useAlarmState();
 
-  /* 编辑态 */
-  const [editingId, setEditingId] = useState<number | null>(null);
-  const [editHour, setEditHour] = useState(0);
-  const [editMinute, setEditMinute] = useState(0);
-  const [editSecond, setEditSecond] = useState(0);
-  const [editLabel, setEditLabel] = useState('');
-  const [editRepeat, setEditRepeat] = useState<Weekday[]>([]);
-  const [editRingtone, setEditRingtone] = useState<SystemAlarmRingtone>(DEFAULT_SYSTEM_ALARM_RINGTONE);
-  const [editLoop, setEditLoop] = useState(true);
+  const {
+    loaded, previewPlaying,
+    editingId,
+    editHour, editMinute, editSecond, editLabel, editRepeat, editRingtone, editLoop,
+    setEditHour, setEditMinute, setEditSecond, setEditLabel, setEditRepeat, setEditRingtone, setEditLoop,
+    adding, setAdding,
+    newHour, newMinute, newSecond, newLabel, newRepeat, newRingtone, newLoop,
+    setNewHour, setNewMinute, setNewSecond, setNewLabel, setNewRepeat, setNewRingtone, setNewLoop,
+    weekdayLabel, repeatSummary, nextRingDesc,
+    addAlarm, closeEditor, deleteAlarm, toggleEnabled, startEdit, saveEdit,
+    sortedAlarms, showEditor,
+  } = state;
 
-  /* 新建态 */
-  const [adding, setAdding] = useState(false);
-  const [newHour, setNewHour] = useState(8);
-  const [newMinute, setNewMinute] = useState(0);
-  const [newSecond, setNewSecond] = useState(0);
-  const [newLabel, setNewLabel] = useState('');
-  const [newRepeat, setNewRepeat] = useState<Weekday[]>([]);
-  const [newRingtone, setNewRingtone] = useState<SystemAlarmRingtone>(DEFAULT_SYSTEM_ALARM_RINGTONE);
-  const [newLoop, setNewLoop] = useState(true);
-
-  /** 启动时从文件加载 */
-  useEffect(() => {
-    let cancelled = false;
-    const applyAlarms = (data: unknown): void => {
-      if (!Array.isArray(data)) return;
-      skipPersistOnceRef.current = true;
-      setAlarms(normalizeAlarms(data as AlarmItem[]));
-    };
-
-    window.api.storeRead(STORE_KEY).then((data) => {
-      if (cancelled) return;
-      if (Array.isArray(data) && data.length > 0) {
-        setAlarms(normalizeAlarms(data as AlarmItem[]));
-      }
-      setLoaded(true);
-    }).catch(() => {
-      if (!cancelled) setLoaded(true);
-    });
-
-    const unsub = window.api.onSettingsChanged((channel: string, value: unknown) => {
-      if (cancelled) return;
-      if (channel === `store:${STORE_KEY}`) {
-        applyAlarms(value);
-      }
-    });
-
-    return () => {
-      cancelled = true;
-      unsub();
-    };
-  }, []);
-
-  /** alarms 变化时持久化 */
-  useEffect(() => {
-    if (!loaded) return;
-    if (skipPersistOnceRef.current) {
-      skipPersistOnceRef.current = false;
-      return;
-    }
-    persistAlarms(alarms);
-  }, [alarms, loaded]);
-
-  useEffect(() => {
-    const unsubscribe = subscribePreviewAlarmSoundState((state) => {
-      setPreviewPlaying(state.playing);
-    });
-    return () => {
-      unsubscribe();
-      stopPreviewAlarmSound();
-    };
-  }, []);
-
-  /** 星期几简写 */
-  const weekdayLabel = useCallback((d: Weekday): string => {
-    const keys = ['sun', 'mon', 'tue', 'wed', 'thu', 'fri', 'sat'];
-    return t(`maxExpand.alarm.weekday.${keys[d]}`, { defaultValue: ['日', '一', '二', '三', '四', '五', '六'][d] });
-  }, [t]);
-
-  /** 重复描述 */
-  const repeatSummary = useCallback((repeat: Weekday[]): string => {
-    if (repeat.length === 0) return t('maxExpand.alarm.repeatOnce', { defaultValue: '仅一次' });
-    if (repeat.length === 7) return t('maxExpand.alarm.repeatEveryday', { defaultValue: '每天' });
-    const weekdays: Weekday[] = [1, 2, 3, 4, 5];
-    const weekend: Weekday[] = [0, 6];
-    if (weekdays.every((d) => repeat.includes(d)) && !weekend.some((d) => repeat.includes(d))) {
-      return t('maxExpand.alarm.repeatWeekdays', { defaultValue: '工作日' });
-    }
-    if (weekend.every((d) => repeat.includes(d)) && !weekdays.some((d) => repeat.includes(d))) {
-      return t('maxExpand.alarm.repeatWeekend', { defaultValue: '周末' });
-    }
-    return repeat.map((d) => weekdayLabel(d)).join(' ');
-  }, [t, weekdayLabel]);
-
-  /** 切换星期 */
-  const toggleWeekday = (list: Weekday[], day: Weekday): Weekday[] => {
-    return list.includes(day) ? list.filter((d) => d !== day) : [...list, day];
-  };
-
-  /** 检查是否已有相同时间的闹钟 */
-  const isDuplicateTime = (h: number, m: number, s: number, excludeId?: number): boolean => {
-    return alarms.some((a) => a.hour === h && a.minute === m && a.second === s && a.id !== excludeId);
-  };
-
-  /** 添加闹钟 */
-  const addAlarm = (): void => {
-    if (isDuplicateTime(newHour, newMinute, newSecond)) return;
-    const item: AlarmItem = {
-      id: Date.now(),
-      hour: newHour,
-      minute: newMinute,
-      second: newSecond,
-      label: newLabel.trim(),
-      enabled: true,
-      repeat: [...newRepeat],
-      ringtone: newRingtone,
-      loop: newLoop,
-      createdAt: Date.now(),
-    };
-    setAlarms((prev) => [...prev, item]);
-    closeEditor();
-  };
-
-  /** 关闭编辑面板（新建 / 编辑共用） */
-  const closeEditor = (): void => {
-    stopPreviewAlarmSound();
-    setAdding(false);
-    setEditingId(null);
-    const _now = new Date();
-    setNewHour(_now.getHours());
-    setNewMinute(_now.getMinutes());
-    setNewSecond(_now.getSeconds());
-    setNewLabel('');
-    setNewRepeat([]);
-    setNewRingtone(DEFAULT_SYSTEM_ALARM_RINGTONE);
-    setNewLoop(true);
-    setEditRingtone(DEFAULT_SYSTEM_ALARM_RINGTONE);
-    setEditLoop(true);
-  };
-
-  /** 删除闹钟 */
-  const deleteAlarm = (id: number): void => {
-    setAlarms((prev) => prev.filter((a) => a.id !== id));
-    if (editingId === id) setEditingId(null);
-  };
-
-  /** 切换开关 */
-  const toggleEnabled = (id: number): void => {
-    setAlarms((prev) => prev.map((a) => a.id === id ? { ...a, enabled: !a.enabled } : a));
-  };
-
-  /** 进入编辑 */
-  const startEdit = (alarm: AlarmItem): void => {
-    setAdding(false);
-    setEditingId(alarm.id);
-    setEditHour(alarm.hour);
-    setEditMinute(alarm.minute);
-    setEditSecond(alarm.second);
-    setEditLabel(alarm.label);
-    setEditRepeat([...alarm.repeat]);
-    setEditRingtone(normalizeSystemAlarmRingtone(alarm.ringtone));
-    setEditLoop(typeof alarm.loop === 'boolean' ? alarm.loop : true);
-  };
-
-  /** 保存编辑 */
-  const saveEdit = (): void => {
-    if (editingId === null) return;
-    if (isDuplicateTime(editHour, editMinute, editSecond, editingId)) return;
-    setAlarms((prev) => prev.map((a) => a.id === editingId ? {
-      ...a,
-      hour: editHour,
-      minute: editMinute,
-      second: editSecond,
-      label: editLabel.trim(),
-      repeat: [...editRepeat],
-      ringtone: editRingtone,
-      loop: editLoop,
-    } : a));
-    closeEditor();
-  };
-
-  /** 计算距下次响铃的时间描述 */
-  const nextRingDesc = useCallback((alarm: AlarmItem): string => {
-    if (!alarm.enabled) return t('maxExpand.alarm.disabled', { defaultValue: '已关闭' });
-    const now = new Date();
-    const todayMinutes = now.getHours() * 60 + now.getMinutes();
-    const alarmMinutes = alarm.hour * 60 + alarm.minute;
-    const todayDay = now.getDay() as Weekday;
-
-    if (alarm.repeat.length === 0) {
-      const diff = alarmMinutes - todayMinutes;
-      if (diff > 0) {
-        const h = Math.floor(diff / 60);
-        const m = diff % 60;
-        return h > 0
-          ? t('maxExpand.alarm.ringIn', { defaultValue: '{{h}}小时{{m}}分钟后', h, m })
-          : t('maxExpand.alarm.ringInMin', { defaultValue: '{{m}}分钟后', m });
-      }
-      return t('maxExpand.alarm.ringTomorrow', { defaultValue: '明天' });
-    }
-
-    for (let offset = 0; offset < 7; offset++) {
-      const checkDay = ((todayDay + offset) % 7) as Weekday;
-      if (!alarm.repeat.includes(checkDay)) continue;
-      if (offset === 0 && alarmMinutes > todayMinutes) {
-        const diff = alarmMinutes - todayMinutes;
-        const h = Math.floor(diff / 60);
-        const m = diff % 60;
-        return h > 0
-          ? t('maxExpand.alarm.ringIn', { defaultValue: '{{h}}小时{{m}}分钟后', h, m })
-          : t('maxExpand.alarm.ringInMin', { defaultValue: '{{m}}分钟后', m });
-      }
-      if (offset === 0) continue;
-      if (offset === 1) return t('maxExpand.alarm.ringTomorrow', { defaultValue: '明天' });
-      return t('maxExpand.alarm.ringInDays', { defaultValue: '{{n}}天后', n: offset });
-    }
-    return '';
-  }, [t]);
-
-  /** 排序：按时间升序 */
-  const sortedAlarms = [...alarms].sort((a, b) => {
-    const ta = a.hour * 60 + a.minute;
-    const tb = b.hour * 60 + b.minute;
-    return ta - tb;
-  });
-
-  /** 是否展示右侧编辑面板 */
-  const showEditor = adding || editingId !== null;
-
-  /** 构建 meta 片段（label · repeat · next），用圆点分隔 */
-  const buildMetaFragments = (alarm: AlarmItem): React.ReactNode[] => {
-    const parts: React.ReactNode[] = [];
-    if (alarm.label) {
-      parts.push(<span key="label" className="alarm-card-label">{alarm.label}</span>);
-    }
-    const rep = repeatSummary(alarm.repeat);
-    if (rep) {
-      if (parts.length > 0) parts.push(<span key="s1" className="alarm-card-meta-sep" />);
-      parts.push(<span key="repeat">{rep}</span>);
-    }
-    const next = nextRingDesc(alarm);
-    if (next) {
-      if (parts.length > 0) parts.push(<span key="s2" className="alarm-card-meta-sep" />);
-      parts.push(<span key="next">{next}</span>);
-    }
-    return parts;
-  };
-
-  /** 当前编辑器用到的时/分/秒/标签/重复 */
+  /** 当前编辑器用到的时/分/秒/标签/重复/铃声/循环 */
   const editorHour = adding ? newHour : editHour;
   const editorMinute = adding ? newMinute : editMinute;
   const editorSecond = adding ? newSecond : editSecond;
@@ -486,7 +79,7 @@ export function AlarmTab(): React.ReactElement {
             type="button"
             onClick={() => {
               if (adding) { closeEditor(); }
-              else { const _now = new Date(); setNewHour(_now.getHours()); setNewMinute(_now.getMinutes()); setNewSecond(_now.getSeconds()); setAdding(true); setEditingId(null); }
+              else { const _now = new Date(); setNewHour(_now.getHours()); setNewMinute(_now.getMinutes()); setNewSecond(_now.getSeconds()); setAdding(true); }
             }}
             title={t('maxExpand.alarm.add', { defaultValue: '新建闹钟' })}
           >
@@ -502,139 +95,45 @@ export function AlarmTab(): React.ReactElement {
             </div>
           )}
           {sortedAlarms.map((alarm) => (
-            <div
+            <AlarmCard
               key={alarm.id}
-              className={`alarm-card${alarm.enabled ? '' : ' alarm-card--disabled'}${editingId === alarm.id ? ' alarm-card--active' : ''}`}
-            >
-              <div className="alarm-card-left" onClick={() => startEdit(alarm)}>
-                <div className="alarm-card-time">{formatTime(alarm.hour, alarm.minute, alarm.second)}</div>
-                <div className="alarm-card-meta">
-                  {buildMetaFragments(alarm)}
-                </div>
-              </div>
-              <div className="alarm-card-right">
-                <button
-                  className="alarm-delete-btn"
-                  type="button"
-                  onClick={() => deleteAlarm(alarm.id)}
-                  title={t('maxExpand.alarm.delete', { defaultValue: '删除' })}
-                >
-                  <img src={SvgIcon.DELETE} alt="" className="alarm-tab-btn-icon" />
-                </button>
-                <button
-                  className={`alarm-toggle${alarm.enabled ? ' alarm-toggle--on' : ''}`}
-                  type="button"
-                  onClick={() => toggleEnabled(alarm.id)}
-                  title={alarm.enabled ? t('maxExpand.alarm.turnOff', { defaultValue: '关闭' }) : t('maxExpand.alarm.turnOn', { defaultValue: '开启' })}
-                >
-                  <span className="alarm-toggle-track">
-                    <span className="alarm-toggle-thumb" />
-                  </span>
-                </button>
-              </div>
-            </div>
+              alarm={alarm}
+              isActive={editingId === alarm.id}
+              weekdayLabel={weekdayLabel}
+              repeatSummary={repeatSummary}
+              nextRingDesc={nextRingDesc}
+              onStartEdit={startEdit}
+              onDelete={deleteAlarm}
+              onToggle={toggleEnabled}
+            />
           ))}
         </div>
       </div>
 
       {/* ── 右侧：编辑面板（始终渲染，通过 CSS 类控制展开） ── */}
-      <div className={`alarm-editor-panel${showEditor ? ' alarm-editor-panel--visible' : ''}`}>
-        {/* 固定顶栏：标题 + 操作按钮 */}
-        <div className="alarm-editor-panel-header">
-          <div className="alarm-editor-panel-title">
-            {adding
-              ? t('maxExpand.alarm.newTitle', { defaultValue: '新建闹钟' })
-              : t('maxExpand.alarm.editTitle', { defaultValue: '编辑闹钟' })}
-          </div>
-          <div className="alarm-editor-actions">
-            <button className="alarm-editor-cancel-btn" type="button" onClick={closeEditor}>
-              {t('maxExpand.alarm.cancel', { defaultValue: '取消' })}
-            </button>
-            <button className="alarm-editor-save-btn" type="button" onClick={adding ? addAlarm : saveEdit}>
-              {t('maxExpand.alarm.save', { defaultValue: '保存' })}
-            </button>
-          </div>
-        </div>
-
-        {/* 可滚动表单区域 */}
-        <div className="alarm-editor-panel-body">
-          <div className="alarm-editor-time-row">
-            <WheelPicker min={0} max={23} value={editorHour} onChange={setEditorHour} />
-            <span className="alarm-editor-time-sep">:</span>
-            <WheelPicker min={0} max={59} value={editorMinute} onChange={setEditorMinute} />
-            <span className="alarm-editor-time-sep">:</span>
-            <WheelPicker min={0} max={59} value={editorSecond} onChange={setEditorSecond} />
-          </div>
-
-          <div className="alarm-editor-field">
-            <div className="alarm-editor-field-label">{t('maxExpand.alarm.labelField', { defaultValue: '备注' })}</div>
-            <input
-              className="alarm-editor-label-input"
-              type="text"
-              placeholder={t('maxExpand.alarm.labelPlaceholder', { defaultValue: '闹钟备注（可选）' })}
-              value={editorLabel}
-              onChange={(e) => setEditorLabel(e.target.value)}
-            />
-          </div>
-
-          <div className="alarm-editor-field">
-            <div className="alarm-editor-field-label-row">
-              <span className="alarm-editor-field-label">{t('maxExpand.alarm.repeatField', { defaultValue: '重复' })}</span>
-              <span className="alarm-editor-repeat-hint">{repeatSummary(editorRepeat)}</span>
-            </div>
-            <div className="alarm-editor-weekdays">
-              {ALL_WEEKDAYS.map((d) => (
-                <button
-                  key={d}
-                  className={`alarm-weekday-btn${editorRepeat.includes(d) ? ' alarm-weekday-btn--active' : ''}`}
-                  type="button"
-                  onClick={() => setEditorRepeat(toggleWeekday(editorRepeat, d))}
-                >
-                  {weekdayLabel(d)}
-                </button>
-              ))}
-            </div>
-          </div>
-
-          <div className="alarm-editor-field">
-            <div className="alarm-editor-field-label-row">
-              <span className="alarm-editor-field-label">{t('maxExpand.alarm.ringtoneField', { defaultValue: '铃声' })}</span>
-              <button
-                className="alarm-editor-preview-btn"
-                type="button"
-                onClick={() => previewAlarmSound(editorRingtone)}
-              >
-                {previewPlaying
-                  ? t('maxExpand.alarm.pausePreviewRingtone', { defaultValue: '暂停' })
-                  : t('maxExpand.alarm.previewRingtone', { defaultValue: '试听' })}
-              </button>
-            </div>
-            <div className="alarm-editor-ringtone-options">
-              {SYSTEM_ALARM_RINGTONE_OPTIONS.map((item) => (
-                <button
-                  key={item.value}
-                  type="button"
-                  className={`alarm-editor-ringtone-btn${editorRingtone === item.value ? ' alarm-editor-ringtone-btn--active' : ''}`}
-                  onClick={() => setEditorRingtone(item.value)}
-                >
-                  {t(item.labelKey, { defaultValue: item.defaultLabel })}
-                </button>
-              ))}
-            </div>
-          </div>
-
-          <div className="alarm-editor-field">
-            <label className="alarm-editor-check-row">
-              <input
-                type="checkbox"
-                checked={editorLoop}
-                onChange={(e) => setEditorLoop(e.target.checked)}
-              />
-              {t('maxExpand.alarm.loopPlayback', { defaultValue: '循环播放铃声' })}
-            </label>
-          </div>
-        </div>
-      </div>
+      <AlarmEditor
+        adding={adding}
+        visible={showEditor}
+        hour={editorHour}
+        minute={editorMinute}
+        second={editorSecond}
+        label={editorLabel}
+        repeat={editorRepeat}
+        ringtone={editorRingtone}
+        loop={editorLoop}
+        previewPlaying={previewPlaying}
+        repeatSummary={repeatSummary}
+        weekdayLabel={weekdayLabel}
+        setHour={setEditorHour}
+        setMinute={setEditorMinute}
+        setSecond={setEditorSecond}
+        setLabel={setEditorLabel}
+        setRepeat={setEditorRepeat}
+        setRingtone={setEditorRingtone}
+        setLoop={setEditorLoop}
+        onCancel={closeEditor}
+        onSave={adding ? addAlarm : saveEdit}
+      />
     </div>
   );
 }
