+++ src/renderer/components/states/maxExpand/components/alarm/config/wheelPickerConstants.ts
@@ -1,0 +1,31 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file wheelPickerConstants.ts
+ * @description 轮盘选择器常量。
+ * @author 鸡哥
+ */
+
+/** 轮盘单项高度 (px) */
+export const WHEEL_ITEM_HEIGHT = 28;
+
+/** 可见项数量 */
+export const WHEEL_VISIBLE_ITEMS = 5;
